refactor to allow for cli or node use
