+++ lib/cheers.js
@@ -51,11 +51,11 @@ var getEmailData = function(body) {
 	return emailData;
 };
 
-module.exports = Cheers = {
-	sendCheers: function(opts) {
+var Cheers = {
+	sendCheers: function(opts, cb) {
 		request(getURI + opts.token, function(err, res, body){
 			if(err) {
-				console.error(err);
+				cb(err);
 			}
 			else {
 				var cheerioBody = cheerio.load(body);
@@ -69,9 +69,14 @@ module.exports = Cheers = {
 				form["respond[cheers][][praise]"] = opts.message;
 
 				request.post({url: postURI + opts.token, form: form}, function (err, res, body) {
-					console.log(err, body);
-					if(body.indexOf(opts.token) !== -1) {
-						console.log("Success!")
+					if(err) {
+						cb(err);
+					}
+					else if(body.indexOf(opts.token) !== -1) {
+						cb(null);
+					}
+					else {
+						cb(new Error("Cheers was unsuccessful. We did not get the expected response."));
 					}
 				});
 			}
@@ -79,7 +84,7 @@ module.exports = Cheers = {
 	},
 
 
-	getCheersPage: function(opts) {
+	getCheersPage: function(opts, cb) {
 		var page = opts.page;
 		var type = opts.type;
 		var token = opts.token;
@@ -91,7 +96,7 @@ module.exports = Cheers = {
 		var cheersURL = "https://www.tinypulse.com/user_portal/cheers/" + cheersURLComponent + "?page=" + page + "&response_token=" + token;
 		request(cheersURL, function(err, res, body){
 			if(err) {
-				console.log(err);
+				cb(err);
 			}
 			else{
 				var $ = cheerio.load(body);
