+++ lib/parseEmailData.js
