+++ lib/token.js
@@ -1,20 +1,20 @@
 var fs = require('fs');
 var mkdirp = require("mkdirp");
-var Browser = require("zombie");
+// var Browser = require("zombie");
 
 
 var appDirectory = process.env.HOME + "/.cheers";
+var dataPath = appDirectory + "/data.json";
 
-module.exports = Token = {
-
-	dataPath : appDirectory + "/data.json",
+var Token = {
+	dataPath: dataPath,
 	update: function(token) {
 		var data = {
 			token: token
 		};
 		
 		mkdirp.sync(appDirectory);
-		fs.writeFile(this.dataPath, JSON.stringify(data), function(err) {
+		fs.writeFile(dataPath, JSON.stringify(data), function(err) {
 			if (err) {
 				console.error(err);
 			}
@@ -47,7 +47,7 @@ module.exports = Token = {
 										var data = {
 											token: token
 										}
-										fs.writeFile(self.dataPath, JSON.stringify(data), function(err) {
+										fs.writeFile(dataPath, JSON.stringify(data), function(err) {
 											if (err) {
 												console.error(err);
 											}
@@ -69,3 +69,5 @@ module.exports = Token = {
 		});
 	}
 }
+
+module.exports = Token;
